pyproject.toml
[project]
name = "valuation-from-spec"
version = "0.1.0"
description = "Portfolio Company Monitoring System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.8.0",
    "pyairtable>=3.2.0",
    "firecrawl-py>=4.5.0",
    "python-dotenv>=1.2.1",
    "structlog>=25.5.0",
    "requests>=2.32.5",
    "anthropic>=0.40.0",
    "Pillow>=12.0.0",
    "imagehash>=4.3.2",
    "beautifulsoup4>=4.14.2",
    "tenacity>=9.1.2",
]

[project.scripts]
airtable-extractor = "src.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.3",
    "mypy>=1.18.2",
]
